Refactor api gw
