+++ .env
@@ -17,3 +17,4 @@ PREFETCH_COUNT=1
 AWS_DEFAULT_REGION=sa-east-1
 AWS_ACCESS_KEY_ID=fakeMyKeyId
 AWS_SECRET_ACCESS_KEY=fakeSecretAccessKey
+BUCKET_NAME=book-recommendations-v2-bucket-sa-east-1
