+++ api-gateway/common/encryptionUtils.go
@@ -1,4 +1,4 @@
-package service
+package common
 
 import (
 	"errors"
